lesson02/requirements_dev.txt
fastavro
flask
requests-mock
pytest